+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Changed
 - return the plugin artifact shrinking with R8 (saved 32.719%, 265.9 KB).
+- update compatibility methods `NamedDomainObjectSet<T>.named*` for Gradle 8.6+ and older.
 
 
 ## [0.7.0-alpha2] - 2024-02-22